src/main/java/music/festival/backend/dto/RecordLabel.java
package music.festival.backend.dto;

import java.util.ArrayList;
import java.util.List;

import org.apache.commons.lang3.builder.EqualsBuilder;
import org.apache.commons.lang3.builder.HashCodeBuilder;
import org.apache.commons.lang3.builder.ToStringBuilder;

public class RecordLabel {

	private String name;

	private List<BandDTO> bands;

	public RecordLabel() {
		bands = new ArrayList<BandDTO>();
	}

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}

	public List<BandDTO> getBands() {
		return bands;
	}

	public void setBands(List<BandDTO> bands) {
		this.bands = bands;
	}

	@Override
	public boolean equals(Object obj) {
		return EqualsBuilder.reflectionEquals(this, obj, true);
	}

	@Override
	public String toString() {
		return ToStringBuilder.reflectionToString(this);
	}

	@Override
	public int hashCode() {
		return HashCodeBuilder.reflectionHashCode(this, true);
	}

}
